backend/requirements.txt
# --- Bestehende Pakete (beibehalten) ---
annotated-types==0.7.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-ai-generativelanguage==0.9.0  # Required by langchain-google-genai; causes warning with google-generativeai but works in practice
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
psycopg==3.2.10  # silvia
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0

# packages for FastAPI + authentification + DB pool
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.9.0
asyncpg==0.30.0

# packages for ETL
langchain==1.0.8
langchain-core==1.1.0
langchain-community==0.4.1
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
langchain-google-genai==2.1.12
pypdf==6.1.3
unstructured==0.18.18
psycopg2-binary==2.9.11
pgvector==0.4.1
bs4==0.0.2
playwright==1.56.0
html2text==2025.4.15
typing==3.7.4.3